LI11718

Description

Tipos de dados e funções auxiliares para a realização do projeto de LI1 em 2017/18.

Tipos de dados

Básicos

type Altura = Int

Altura de uma peça num Tabuleiro.

data Orientacao

Orientações no mapa.

Constructors

Norte

Este

Sul

Oeste

type Posicao = (Int, Int)

Posições num Tabuleiro.

type Ponto = (Double, Double)

Ponto no mapa.

type Angulo = Double

Ângulo em graus.

type Tempo = Double

Períodos de tempo.

type Velocidade = Ponto

Vectores de velocidade.

Caminhos

type Caminho = [Passo]

Uma sequência de passos que dá origem a um mapa.

data Passo

Tipos de passos que podem occorrer num Caminho.

Constructors

Avanca
Segue em frente e mantem o nível

Sobe
Segue em frente e sobe um nível

Desce
Segue em frente e desce um nível

CurvaEsq
Curva à esquerda e mantem o nível

CurvaDir
Curva à direita e mantem o nível

Mapas

data Mapa

Mapa de um jogo, composto por um tabuleiro, uma posição inicial e uma altura inicial.

Constructors

Mapa (Posicao, Orientacao) Tabuleiro

type Tabuleiro = [[Peca]]

O tabuleiro do mapa, representado por uma matriz de Peca.

data Peca

Uma peça num Tabuleiro. A altura atribuída é sempre a do ponto mais baixo da peça.

Constructors

Peca Tipo Altura

data Tipo

Tipos de peças contidos num Tabuleiro.

Constructors

Rampa Orientacao

Curva Orientacao

Recta

Lava

type Dimensao = (Int, Int)

Dimensões de um Tabuleiro.

data Carro

O estado de um carro dentro de um Mapa. A direção da velocidade (movimento) não é necessariamente a mesma da direção do carro.

Constructors

Carro

Fields

posicao :: Ponto

a posição atual do carro

direcao :: Angulo

a direção atual do carro

velocidade :: Velocidade

a velocidade atual do carro

Jogo

data Jogo

O estado de um jogo dentro, incluindo o mapa e as suas propriedades e o estado dos vários jogadores. Os campos carros, nitros e historico devem ter a mesma dimensão, o número de jogadores.

Constructors

Jogo

Fields

mapa :: Mapa

o mapa do percurso

pista :: Propriedades

as propriedades do percurso

carros :: [Carro]

o estado do carro de cada jogador

nitros :: [Tempo]

a quantidade de nitro disponível para cada jogador

historico :: [[Posicao]]

o histórico de posições de cada jogador

data Propriedades

As propriedades físicas de um percurso.

Constructors

Propriedades

Fields

k_atrito :: Double

o atrito do piso

k_pneus :: Double

o atrito dos pneus

k_acel :: Double

a intensidade da aceleração

k_peso :: Double

o peso do carro

k_nitro :: Double

a intensidade do nitro

k_roda :: Double

a sensibilidade do guiador

data Acao

As ações que podem ser tomadas por cada jogador em cada instante.

Constructors

Acao

Fields

acelerar :: Bool

se está a acelerar

travar :: Bool

se está a travar

esquerda :: Bool

se está a curvar para a esquerda

direita :: Bool

se está a curvar para a direita

nitro :: Maybe Int

se o nitro está ativo em algum jogador

Funções auxiliares fornecidas

dimensao :: Caminho -> Dimensao

Dado um caminho, calcula a dimensão do tabuleiro correspondente.

>>> dimensao [CurvaDir,CurvaDir,CurvaDir,CurvaDir]
(4,4)

partida :: Caminho -> Posicao

Dado um caminho, calcula a sua posição inicial no tabuleiro correspondente.

>>> partida [CurvaDir,CurvaDir,CurvaDir,CurvaDir]
(2,1)

dirInit :: Orientacao

Direção inicial na construção de um caminho.

dirInit == Este

altInit :: Altura

Altura inicial na construção de um caminho.

altInit == 0

altLava :: Altura

Altura da lava.

altLava == 0