Main

Description

This module presents a minimalistic example of the usage of HAAP, consisting on testing an HSpec specification, running several code analysis plugins, providing global feedback with the ranking and tournament plugins, visualization of solutions with CodeWorld, all connected by a webpage generated by Hakyll.

Documentation

example :: Project

emptyLi1DB :: Example_DB

data Example_DB

Constructors

Example_DB

Fields

exTourneyDB :: HaapTourneyDB ExPlayer

lnsTourney :: DBLens (BinaryDB Example_DB) (HaapTourneyDB ExPlayer)

main :: IO ((), HaapLog)

An HAAP script that runs tests with the HSpec plugin, runs several code analysis plugins, generates feedback with the ranking and tournament plugins, provides visualization with the CodeWorld plugin and uses the Hakyll plugin to generate a webpage that connects all artifacts.

exSpec :: HaapSpec

A concrete specification to be tested.

exRankSpec :: Int -> HaapSpec

A specification to be used in a ranking, will be run for each group.

exRankScore :: HaapTestRes -> FloatScore

Calculates a quantitative score from the results of the tests.

exRank :: HaapStack t m => HaapSpecRank t m Int FloatScore

Builds a global ranking according to the result of exRankSpec for each group, normalized according to exRankScore.

data ExPlayer

Constructors

ExPlayer (String, Bool)

exTourney :: MonadIO m => HaapTourney t m (BinaryDB Example_DB) ExPlayer Link

codeworldDrawArgs :: CodeWorldArgs

Arguments for the CodeWorld plugin to draw maps.

codeworldGameArgs :: CodeWorldArgs

Arguments for the CodeWorld plugin to play a game.