Mapas

Documentation

all_mapas :: [Mapa]

mapas_torneio :: [Mapa]

mapas :: [[Mapa]]

mapas128 :: [Mapa]

Mapas predefinidos utilizados no torneio dos bots na ronda de 128. Apenas mapas com alturas negativas, sem rampas, sem curvas em U, sem peças adjacentes (sem possibilidade de atalhos) e sem cruzamentos.

mapas64 :: [Mapa]

Mapas predefinidos utilizados no torneio dos bots na ronda de 64. Igual ao mapas128 mas agora com alturas positivas, rampas pontuais e curvas em U.

mapas16 :: [Mapa]

Mapas predefinidos utilizados no torneio dos bots na ronda de 16. Igual ao mapas64 mas agora com peças adjacentes e cruzamentos pontuais.

mapas4 :: [Mapa]

Mapas predefinidos utilizados no torneio dos bots na ronda de 4, a final. Vale tudo.

propriedades :: [Propriedades]

Propriedades do terreno predefinidas a ser utilizadas no torneio dos bots.

terra :: Propriedades

gelo :: Propriedades

asfalto :: Propriedades

bracketsFinal :: [[Int]]

brackets_final :: [[Int]]