Tarefa6_2017li1g183

Description

Módulo para a realização da Tarefa 6 de LI1 em 2017/18.

Documentation

bot

Arguments

:: Tempo
tempo decorrido desde a última decisão

-> Jogo
estado atual do jogo

-> Int
identificador do jogador dentro do estado

-> Acao
a decisão tomada pelo bot

Função usada para simular um bot no jogo Micro Machines. Em cada instante, dado o tempo decorrido, o estado do jogo e o identificador do jogador, toma uma ação.